create classe usecase services folder and move hooks to it
