+++ src/features/classe/application/repository/classe-repository.ts
@@ -2,10 +2,10 @@ import useCreateClasseInfra from "@/features/classe/infra/services/useCreateClas
 import useDeleteClasseInfra from "@/features/classe/infra/services/useDeleteClasseInfra";
 import useGetClasseInfra from "@/features/classe/infra/services/useGetClasseInfra";
 import useGetClassesInfra from "@/features/classe/infra/services/useGetClassesInfra";
-import useCreateClasseUsecase from "../usecases/useCreateClasseUsecase";
-import useDeleteClasseUsecase from "../usecases/useDeleteClasseUsecase";
-import useGetClasseUsecase from "../usecases/useGetClasseUsecase";
-import useGetClassesUsecase from "../usecases/useGetClassesUsecase";
+import useCreateClasseUsecase from "../usecases/services/useCreateClasseUsecase";
+import useDeleteClasseUsecase from "../usecases/services/useDeleteClasseUsecase";
+import useGetClasseUsecase from "../usecases/services/useGetClasseUsecase";
+import useGetClassesUsecase from "../usecases/services/useGetClassesUsecase";
 import authRepository from "@/features/auth/application/repository/old-auth-repository";
 
 export const classeRepository = {
