+++ src/features/classe/application/usecases/index.ts
@@ -1,5 +1,1 @@
-export * from './classe-usecases';
-export * from './useCreateClasseUsecase';
-export * from './useDeleteClasseUsecase';
-export * from './useGetClasseUsecase';
-export * from './useGetClassesUsecase';
+export * from "./classe-usecases";
