+++ src/features/classe/application/usecases/services/useUpdateClasseVisibility.ts
@@ -1,0 +1,5 @@
+import React from "react";
+
+function useUpdateClasseVisibility() {}
+
+export default useUpdateClasseVisibility;
